Initial setup and structure
Primary concept is to toggle specific CSS rules to override the default styling of the
player page. Via a worker, we're storing which tabs have inserted CSS overrides and
responding by inserting or removing the stylesheet as well as toggling the extension
icon.

Additionally, we're listening to active tab changes and removals to keep memory low, but
account for multiple f1tv pages open simultaneously.

Potential improvements including a move to TS, including Chrome extension types, and refactoring
the worker file with a functional approach.